requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0 
idna==3.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
python-multipart==0.0.9
email_validator==2.2.0
passlib==1.7.4
httpx==0.27.2
asyncio==3.4.3
pytz==2024.2
boto3==1.35.57